requirements.txt
paho-mqtt
pyinstaller
PySide6
zeroconf
numpy
pyserial